+++ src/test/java/apiTestScenarios/CreateUserWithPostApiAuthRegisterTest.java
@@ -9,21 +9,16 @@ import org.junit.Before;
 import org.junit.Test;
 import serialsClasses.ApiEndPoints;
 import serialsClasses.UserSerials;
-import java.util.ArrayList;
-import java.util.List;
-
 import static io.restassured.RestAssured.given;
 import static org.hamcrest.CoreMatchers.containsString;
 import static org.hamcrest.CoreMatchers.equalTo;
 
-public class GetUsersOrdersWithGetApiOrders {
-
+public class CreateUserWithPostApiAuthRegisterTest {
     UserSerials user;
     ApiEndPoints api = new ApiEndPoints();
     String email;
     String password;
     String name;
-    List<String> ingredients = new ArrayList<>();
     Response response;
     String token = "";
 
@@ -33,46 +28,17 @@ public class GetUsersOrdersWithGetApiOrders {
         email = faker.internet().emailAddress();
         password = faker.internet().password();
         name = faker.name().firstName();
-        ingredients.add("61c0c5a71d1f82001bdaaa6d");
-        ingredients.add("61c0c5a71d1f82001bdaaa6f");
         RestAssured.baseURI = "https://stellarburgers.nomoreparties.site";
-        response = sendRequest(email, password, name);
-        token = getToken(response).replaceFirst("Bearer ", "");
-        sendPostOrderRequest(token, ingredients);
     }
 
-    @Step("Send POST request")
+    @Step ("Send POST request")
     public Response sendRequest(String email, String password, String name){
         user = new UserSerials(email, password, name);
-
         response = given()
                 .header("Content-type", "application/json")
                 .body(user)
                 .when()
                 .post(api.getCreateUser());
-        return response;
-    }
-
-    @Step("Send POST order request")
-    public Response sendPostOrderRequest(String token, List<String> ingredients){
-        user = new UserSerials(ingredients);
-
-        response = given()
-                .header("Content-type", "application/json")
-                .auth().oauth2(token)
-                .and()
-                .body(user)
-                .when()
-                .post(api.getCreateOrder());
-        return response;
-    }
-
-    @Step("Send GET order request")
-    public Response sendGetOrderRequest(String token){
-        response = given()
-                .header("Content-type", "application/json")
-                .auth().oauth2(token)
-                .get(api.getCreateOrder());
         return response;
     }
 
@@ -85,25 +51,51 @@ public class GetUsersOrdersWithGetApiOrders {
     }
 
     @Step ("Get Token")
-    public String getToken(Response response) {
+    public String getToken(Response response){
         return response.then().extract().body().path("accessToken");
     }
 
     @Test
-    public void canGetOrdersWithAuthUser(){
-        response = sendGetOrderRequest(token);
+    public void canCreateUser(){
+        response = sendRequest(email, password, name);
+
         checkStatus (response, true, "200 OK");
     }
 
     @Test
-    public void cannotGetOrdersWithoutAuthUser(){
-        response = sendGetOrderRequest("");
-        checkStatus (response, false, "401 Unauthorized");
+    public void cannotCreateTwoIdenticalUsers(){
+        response = sendRequest(email, password, name);
+        Response response1 = sendRequest(email, password, name);
+
+        checkStatus (response1, false, "403 Forbidden");
+
+    }
+
+    @Test
+    public void cannotCreateUserWithoutEmail(){
+        response = sendRequest("", password, name);
+
+        checkStatus (response, false, "403 Forbidden");
+    }
+
+    @Test
+    public void cannotCreateUserWithoutPassword(){
+        response = sendRequest(email, "", name);
+
+        checkStatus (response, false,"403 Forbidden");
+    }
+
+    @Test
+    public void cannotCreateUserWithoutName(){
+        response = sendRequest(email, password, "");
+
+        checkStatus (response, false,"403 Forbidden");
     }
 
     @After
     public void clearAll(){
         try {
+            token = getToken(response).replaceFirst("Bearer ", "");
             given()
                     .auth().oauth2(token)
                     .delete(api.getChangeUser());
@@ -111,4 +103,5 @@ public class GetUsersOrdersWithGetApiOrders {
             token = "";
         }
     }
+
 }
