+++ src/test/java/apiTestScenarios/LoginUserWithPostApiAuthLoginTest.java
@@ -14,7 +14,7 @@ import static io.restassured.RestAssured.given;
 import static org.hamcrest.CoreMatchers.containsString;
 import static org.hamcrest.CoreMatchers.equalTo;
 
-public class LoginUserWithPostApiAuthLogin {
+public class LoginUserWithPostApiAuthLoginTest {
     UserSerials user;
     ApiEndPoints api = new ApiEndPoints();
     String email;
